Home Screen is now fully ready

+++ app/src/main/java/com/potaninpm/soundr/domain/model/UserInfo.kt
@@ -4,7 +4,7 @@ data class UserInfo(
     val name: String,
     val streak: Int,
     val bestStreak: Int,
-    val totalTrainings: Int,
-    val totalTrainingsTime: Int,
+    val totalTrainings: Long,
+    val totalTrainingsTime: Long,
     val progress: Float
 )
